refactor: usePostReplyのリファクタ
